Move db test helper func to dbTestHelpers.ts
